Limits the number of rescaling checks to 100
Currently, every call to rescale a value from the prior checks if the
input is valid. For use cases where the likelihood evaluation takes
longer than ~1e-2 seconds, the overhead is negligible. But, once the
likelihood evaluation is shorter than this, it can be come noticable
while profiling. This change limits the number of checks to 100 for each
prior before the test is skipped.